third_party/blink/web_tests/fast/js/script-line-number-expected.txt

This test checks that line numbers are correctly reported for both inline scripts and inline event handlers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 13 is 13
PASS 16 is 16
PASS successfullyParsed is true

TEST COMPLETE
